用户表添加主键使用序列值

+++ src/main/java/com/bd/service/UserService.java
@@ -72,5 +72,5 @@ public interface UserService {
 	 * @param userNum
 	 * @return
 	 */
-	Long batchInsertUserInfo(Long userNum);
+	Integer batchInsertUserInfo(Long userNum);
 }
